fix: proxy_cli calling the wrong binary
In overrides, the way fuelup proxies a specific version of a component is
different from the way it proxies a general toolchain. Components in a
toolchain override can be executed via the toolchain directory, but if
a specific version is declared, we have to call it from the store,
which means there must be some logic that derives the binary path.

Previously, the cod erroneously derived the same path for both situations.
When an entire toolchain is overridden, we can just call 'proc_name' directly.