fix: linking interaction with conflict update tests
We were creating manually creating some binaries in the UpdateConflict
state during update tests. This removes the manual creation in favour
of linking the binaries instead (which is what actual fuelup does anyway),
then removing some of the linked binaries in that configuration in order
to trigger certain warning messages.