fix: proxy_cli erroneously executes `forc` when forc plugins are called (#423)
* fix: proxy_cli calling the wrong binary

In overrides, the way fuelup proxies a specific version of a component is
different from the way it proxies a general toolchain. Components in a
toolchain override can be executed via the toolchain directory, but if
a specific version is declared, we have to call it from the store,
which means there must be some logic that derives the binary path.

Previously, the cod erroneously derived the same path for both situations.
When an entire toolchain is overridden, we can just call 'proc_name' directly.

* expose hard_or_symlink_file for tests

* test: add tests to check if correct binaries are called post override

* fix: linking interaction with conflict update tests

We were creating manually creating some binaries in the UpdateConflict
state during update tests. This removes the manual creation in favour
of linking the binaries instead (which is what actual fuelup does anyway),
then removing some of the linked binaries in that configuration in order
to trigger certain warning messages.

* doc: add comments to explain test

* add doc comments; make cfg.fuelup() use cfg.exec(); set CARGO_HOME in exec()