chore: deprecate usage of hashes to check download duplication (#431)
Prior to the 'store' model that we now use, each binary was not cached
in some local location. We used basic sha256 hashes stored within .fuelup/hashes
directory as a way to tell fuelup that a toolchain already exists. This is important
to ensure that users don't waste time re-downloading binaries that they have
already downloaded once before.

With the store model, we no longer need this since it's much faster now
to re-install toolchains with duplicate binaries since they are all cached
in the store. We can now get rid of this hacky way of speeding up the download
process, since it's built-in to the design.